daterangepickerOptions

Update the daterangepicker

Usage

daterangepickerOptions(
  minYear = NULL,
  maxYear = NULL,
  showDropdowns = TRUE,
  showCustomRangeLabel = TRUE,
  opens = c("right", "left", "center"),
  drops = c("down", "up"),
  timePicker = FALSE,
  timePickerIncrement = 1,
  timePicker24Hour = FALSE,
  timePickerSeconds = FALSE,
  showWeekNumbers = FALSE,
  showISOWeekNumbers = FALSE,
  parentEl = NULL,
  maxSpan = NULL,
  alwaysShowCalendars = FALSE,
  buttonClasses = NULL,
  applyButtonClasses = NULL,
  cancelButtonClasses = NULL,
  cancelIsClear = FALSE,
  locale = NULL,
  singleDatePicker = FALSE,
  autoApply = FALSE,
  linkedCalendars = TRUE,
  autoUpdateInput = TRUE,
  ...
)

Arguments

minYear

The minimum year as numeric shown in the dropdowns when showDropdowns is set to TRUE

maxYear

The maximum year as numeric shown in the dropdowns when showDropdowns is set to TRUE

showDropdowns

Show year and month select boxes above calendars to jump to a specific month and year.

showCustomRangeLabel

Displays "Custom Range" at the end of the list of predefined ranges, when the ranges option is used. This option will be highlighted whenever the current date range selection does not match one of the predefined ranges. Clicking it will display the calendars to select a new range.

opens

Whether the picker appears aligned to the left, to the right, or to the center under the HTML element it's attached to.

drops

Whether the picker appears below (down) or above (up) the HTML element it's attached to. Default is down

timePicker

Adds select boxes to choose times in addition to dates.

timePickerIncrement

Increment of the minutes selection list for times (i.e. 30 to allow only selection of times ending in 0 or 30).

timePicker24Hour

Use 24-hour instead of 12-hour times, removing the AM/PM selection.

timePickerSeconds

Show seconds in the timePicker.

showWeekNumbers

Show localized week numbers at the start of each week on the calendars.

showISOWeekNumbers

Show ISO week numbers at the start of each week on the calendars.

parentEl

jQuery selector of the parent element that the date range picker will be added to, if not provided this will be 'body'

maxSpan

The maximum span between the selected start and end dates as List or table. You can provide any object the moment library would let you add to a date.

alwaysShowCalendars

Normally, if you use the ranges option to specify pre-defined date ranges, calendars for choosing a custom date range are not shown until the user clicks "Custom Range". When this option is set to TRUE, the calendars for choosing a custom date range are always shown instead.

buttonClasses

CSS class names that will be added to both the apply and cancel buttons.

applyButtonClasses

CSS class names that will be added only to the apply button.

cancelButtonClasses

CSS class names that will be added only to the cancel button.

cancelIsClear

If TRUE, will treat the Cancel Button like a Clear Button.

locale

Allows you to provide localized strings for buttons and labels, customize the date format, and change the first day of week for the calendars. See the examples in `./inst/examples/`

singleDatePicker

Show only a single calendar to choose one date, instead of a range picker with two calendars. The start and end dates provided to your callback will be the same single date chosen.

autoApply

Hide the apply and cancel buttons, and automatically apply a new date range as soon as two dates are clicked.

linkedCalendars

When enabled, the two calendars displayed will always be for two sequential months (i.e. January and February), and both will be advanced when clicking the left or right arrows above the calendars. When disabled, the two calendars can be individually advanced and display any month/year.

autoUpdateInput

Indicates whether the date range picker should automatically update the value of the <input> element it's attached to at initialization and when the selected dates change.

...

Further arguments passed to `daterangepicker`, like isInvalidDate or isCustomDate.

See also